+++ Addon/ElvUI_mMediaTag/core/changelog/3.50.lua
@@ -1,0 +1,9 @@
+mMT.Changelog[350] = {
+	DATE = "08.08.2024",
+	UPDATE = {
+		"[Portrait]: improved event Handler.",
+	},
+	FIX = {
+		"[Portrait]: the wrong portraits were displayed for Group frames.",
+	},
+}