+++ Addon/ElvUI_mMediaTag/core/options/changelog.lua
@@ -1,176 +1,159 @@
 local E = unpack(ElvUI)
 local L = mMT.Locales
 
-local _G = _G
 local tinsert = tinsert
-local change_log_important_string, change_log_new_string, change_log_update_string, change_log_fix_string = nil, nil, nil, nil
-local green, blue, yellow, red, endtag = "|CFF00D80E", "|CFF00A9FF", "|CFFFFCC00", "|CFFFF0048", "|r"
-local new, fix, update, important, dash = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\misc\\star.tga:14:14|t", "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\misc\\done1.tga:14:14|t", "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\datatext\\upgrade7.tga:14:14|t", "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\tags\\quest2.tga:14:14|t", "> "
-
-local change_log_important = {
-	red .. "!!! Important" .. endtag,
-	"",
-	"Dock Guild & Friends drop down menu for Invite and Whisper can be buggy, this is known and will be fixed later.",
-}
-
-local releasdate = "08.08.2024"
-
-local change_log_new = {
-	"Objective Tracker: can now show Quest Level",
-	"Objective Tracker: auto track/ untrack Quest if they are on map",
-}
-
-local change_log_update = {
-	"Objective Tracker Background color function/ behavior",
-	"Improved Portraits code",
-}
-
-local change_log_fix = {
-	"Taints with the Objective Tracker Skin.",
-	"Nil error with Objective Tracker: Achievements and Professions",
-	"Nil error with Objective Tracker Skin in Dungeons.",
-	"Nil error with TAG targeting Icons.",
-	"Portraits: Rare/ Elite Border and Custom colors of MerathilisUI & Eltruism",
-	"Portraits: Texture Coordinates with JiberishUI Icons",
-	"Portraits: the wrong portraits were displayed for Group frames.",
-	"Portraits: Boss Portraits cast icon toggle.",
-}
-
-local function Concatenation(tbl, icon, color)
-	local string = ""
-	for _, line in pairs(tbl) do
-		if color then
-			line = color .. line .. endtag
-		end
-
-		if icon then
-			line = icon .. "  " .. line
-		end
-
-		string = string .. ((line ~= "") and dash or "") .. "  " .. line .. "\n"
+local dash = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\system\\dash.tga:14:14|t  "
+local function Color(string)
+	if type(string) ~= "string" then
+		string = tostring(string)
 	end
-	return string
+
+	return "[" .. "|CFF0393FF" .. string .. "|r" .. "]"
+end
+
+local function renderChangeLogLine(line)
+	line = gsub(line, "%[([^%[]+)%]", Color)
+	return line
 end
 
 local function configTable()
-	--change_log_important_string = Concatenation(change_log_important)
-	change_log_new_string = Concatenation(change_log_new)
-	change_log_update_string = Concatenation(change_log_update)
-	change_log_fix_string = Concatenation(change_log_fix)
-	E.Options.args.mMT.args.changelog.args = {
-		header_changelog = {
-			order = 1,
+	E.Options.args.mMT.args.changelog.args = {}
+
+	for version, data in pairs(mMT.Changelog) do
+		local versionString = format("%d.%02d", version / 100, mod(tonumber(version), 100))
+
+		E.Options.args.mMT.args.changelog.args[tostring(version)] = {
+			order = 1000 - version,
 			type = "group",
-			inline = true,
-			name = mMT.IconSquare .. "  " .. mMT.Name .. "  " .. L["Changelog"],
-			args = {
-				header_version = {
-					order = 1,
-					type = "group",
-					inline = true,
-					name = L["Version:"],
-					args = {
-						version = {
-							order = 1,
-							type = "description",
-							fontSize = "medium",
-							name = green .. mMT.Version .. endtag,
-						},
-						date = {
-							order = 2,
-							type = "description",
-							fontSize = "medium",
-							name = green .. L["Release"] .. endtag .. " " .. (releasdate or "ERROR"),
-						},
+			guiInline = false,
+			name = versionString,
+			args = {},
+		}
+
+		local page = E.Options.args.mMT.args.changelog.args[tostring(version)].args
+
+		page.date = {
+			order = 1,
+			type = "description",
+			name = "|CFFBBBBBB" .. data.DATE
+			.. " " .. L["Released"] .. "|r",
+			fontSize = "small",
+		}
+
+		page.version = {
+			order = 2,
+			type = "description",
+			name = L["Version"] .. " |CFF99FF33" .. versionString .. "|r",
+			fontSize = "large",
+		}
+
+		page.space = {
+			order = 3,
+			type = "description",
+			name = "\n\n",
+			fontSize = "medium",
+		}
+
+		local importantPart = data and data.IMPORTANT
+		if importantPart and #importantPart > 0 then
+			local iconImportant = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\tags\\quest2.tga:14:14|t"
+			page.important = {
+				order = 4,
+				type = "group",
+				guiInline = true,
+				name = iconImportant .. "  |CFFFF006C" .. L["Important"] .. "|r",
+				args = {
+					text = {
+						order = 1,
+						type = "description",
+						fontSize = "medium",
+						name = function()
+							local text = ""
+							for _, line in ipairs(importantPart) do
+								text = text .. dash .. renderChangeLogLine(line) .. "\n"
+							end
+							return text .. "\n"
+						end,
 					},
 				},
-				header_important = {
-					order = 2,
-					type = "group",
-					inline = true,
-					name = important .. "  " .. red .. L["Important:"] .. endtag,
-					hidden = function()
-						if change_log_important_string then
-							return false
-						else
-							return true
-						end
-					end,
-					args = {
-						important = {
-							order = 1,
-							type = "description",
-							fontSize = "large",
-							name = change_log_important_string or "",
-						},
+			}
+		end
+
+		local newPart = data and data.NEW
+		if newPart and #newPart > 0 then
+			local iconNew = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\misc\\star.tga:14:14|t"
+			page.new = {
+				order = 5,
+				type = "group",
+				guiInline = true,
+				name = iconNew .. "  |CFF6559F1" .. L["New"] .. "|r",
+				args = {
+					text = {
+						order = 1,
+						type = "description",
+						fontSize = "medium",
+						name = function()
+							local text = ""
+							for _, line in ipairs(newPart) do
+								text = text .. dash .. renderChangeLogLine(line) .. "\n"
+							end
+							return text .. "\n"
+						end,
 					},
 				},
-				header_new = {
-					order = 3,
-					type = "group",
-					inline = true,
-					name = new .. "  " .. green .. L["New:"] .. endtag,
-					hidden = function()
-						if change_log_new_string then
-							return false
-						else
-							return true
-						end
-					end,
-					args = {
-						new = {
-							order = 1,
-							type = "description",
-							fontSize = "medium",
-							name = change_log_new_string or "",
-						},
+			}
+		end
+
+		local updatePart = data and data.UPDATE
+		if updatePart and #updatePart > 0 then
+			local iconUpdate = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\datatext\\upgrade7.tga:14:14|t"
+			page.update = {
+				order = 6,
+				type = "group",
+				guiInline = true,
+				name = iconUpdate .. "  |CFFBC26E5" .. L["Updates"] .. "|r",
+				args = {
+					text = {
+						order = 1,
+						type = "description",
+						fontSize = "medium",
+						name = function()
+							local text = ""
+							for _, line in ipairs(updatePart) do
+								text = text .. dash .. renderChangeLogLine(line) .. "\n"
+							end
+							return text .. "\n"
+						end,
 					},
 				},
-				header_update = {
-					order = 4,
-					type = "group",
-					inline = true,
-					name = update .. "  " .. blue .. L["Update:"] .. endtag,
-					hidden = function()
-						if change_log_update_string then
-							return false
-						else
-							return true
-						end
-					end,
-					args = {
-						update = {
-							order = 1,
-							type = "description",
-							fontSize = "medium",
-							name = change_log_update_string or "",
-						},
+			}
+		end
+
+		local fixPart = data and data.FIX
+		if fixPart and #fixPart > 0 then
+			local iconFixe = "|TInterface\\Addons\\ElvUI_mMediaTag\\media\\icons\\misc\\done1.tga:14:14|t"
+			page.fix = {
+				order = 7,
+				type = "group",
+				guiInline = true,
+				name = iconFixe .. "  |CFFFFA800" .. L["Fixes"] .. "|r",
+				args = {
+					text = {
+						order = 1,
+						type = "description",
+						fontSize = "medium",
+						name = function()
+							local text = ""
+							for _, line in ipairs(fixPart) do
+								text = text .. dash .. renderChangeLogLine(line) .. "\n"
+							end
+							return text .. "\n"
+						end,
 					},
 				},
-				header_fix = {
-					order = 5,
-					type = "group",
-					inline = true,
-					name = fix .. "  " .. yellow .. L["Fix:"] .. endtag,
-					hidden = function()
-						if change_log_fix_string then
-							return false
-						else
-							return true
-						end
-					end,
-					args = {
-						fix = {
-							order = 1,
-							type = "description",
-							fontSize = "medium",
-							name = change_log_fix_string or "",
-						},
-					},
-				},
-			},
-		},
-	}
+			}
+		end
+	end
 end
 
 tinsert(mMT.Config, configTable)