+++ Addon/ElvUI_mMediaTag/modules/datatexts/Dungeon.lua
@@ -122,4 +122,4 @@ local function OnLeave(self)
 	DT.tooltip:Hide()
 end
 
-DT:RegisterDatatext("mDungeon", "mMediaTag", { "CHALLENGE_MODE_START", "CHALLENGE_MODE_COMPLETED", "PLAYER_ENTERING_WORLD", "UPDATE_INSTANCE_INFO" }, OnEvent, nil, OnClick, OnEnter, OnLeave, mText, nil, nil)
+DT:RegisterDatatext("mDungeon", mMT.DatatextString, { "CHALLENGE_MODE_START", "CHALLENGE_MODE_COMPLETED", "PLAYER_ENTERING_WORLD", "UPDATE_INSTANCE_INFO" }, OnEvent, nil, OnClick, OnEnter, OnLeave, mText, nil, nil)